quantum-shield/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "quantum-shield"
version = "0.1.0"
description = "Quantum-secure cryptographic library with hybrid post-quantum encryption"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    { name = "Tushar Agrawal" }
]
keywords = ["cryptography", "post-quantum", "encryption", "security", "pqc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "cryptography>=41.0",
    "argon2-cffi>=23.1",
]

[project.optional-dependencies]
pq = [
    "oqs>=0.9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
]

[project.urls]
Homepage = "https://github.com/tushar-agrawal/quantum-shield"
Documentation = "https://github.com/tushar-agrawal/quantum-shield#readme"
Repository = "https://github.com/tushar-agrawal/quantum-shield"

[tool.setuptools.packages.find]
include = ["quantum_shield*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
